Removed unmecessary files
